App/requirements.txt
streamlit
torch
torchvision
Pillow
wikipedia
